+++ docs-src/pages/changelog.svelte
@@ -2,13 +2,17 @@
 <h2>v8.0.0 <em>(2023-07-?)</em></h2>
 <ul>
 <li><strong>New:</strong> <code>Label</code> component.</li>
+<li><strong>New icons:</strong> <code>sun</code> and <code>moon</code> for the dark-theme switchers.</li>
 <li><strong>Improved:</strong> <code>info</code>, <code>error</code> and <code>label</code> attributes are now supported on other inputs (<code>Autocomplete</code>, <code>InputDate</code>, <code>Select</code>, <code>ButtonToggle</code>, and <code>Toggle</code>).</li>
+<li>Added <code>title</code> attribute to <code>ButtonToggle</code>.</li>
+<li>Fixed <code>selectable=false</code> not working on <code>Table</code>.</li>
 </ul>
 <h3>Breaking changes</h3>
 <ul>
 <li><code>Datepicker</code> was renamed to <code>InputDate</code>.</li>
 <li><code>Autocomplete</code> - HTML structure changed: <code>.autocomplete input</code> --&gt; <code>.autocomplete .input-inner .input-row input</code></li>
 <li><code>Select</code> - HTML structure changed: <code>.select-wrap select</code> --&gt; <code>.select .input-inner .input-row select</code></li>
+<li><code>Table</code> - CSS classes changed from <code>.table-wrapper table.table</code> --&gt; <code>.table table</code></li>
 <li><code>Toggle</code> - HTML structure changed from <code>.toggle .toggle-inner .toggle-scroller input</code> --&gt; <code>.toggle .toggle-inner .toggle-label .toggle-scroller input</code></li>
 </ul>
 <hr>