+++ src/table/Table.svelte
@@ -1,14 +1,14 @@
 <div
-	class="table-wrapper {className}"
+	class="table {className}"
 	class:round
-	class:selectable
+	class:selectable="{_selectable}"
 	bind:this="{_this}"
 	on:click="{onClick}"
 	on:focus|capture="{onFocus}"
 	on:keydown="{onKeyDown}"
 	on:dblclick="{onDblClick}">
 
-	<table class="table"><slot /></table>
+	<table><slot /></table>
 </div>
 
 <script>
@@ -34,10 +34,12 @@ let headerHeight = 0;
 let clickTimer;
 let previousKey;
 
+$:_selectable = (selectable === true || selectable === 'true');
+
 
 onMount(() => {
 	Object.assign(_this.dataset, data);
-	if (selectable === true || selectable === 'true') {
+	if (_selectable) {
 		makeRowsSelectable();
 		requestAnimationFrame(() => {
 			const head = _this && _this.querySelector('thead');
@@ -48,7 +50,7 @@ onMount(() => {
 
 
 onDestroy(() => {
-	if (selectable === true || selectable === 'true') makeRowsNotSelectable();
+	if (_selectable) makeRowsNotSelectable();
 });
 
 
@@ -140,7 +142,7 @@ function selectFocusedRow (rowEl) {
 
 
 function onFocus (e) {
-	if (selectable !== true && selectable !== 'true') return;
+	if (!_selectable) return;
 	if (!_this.contains(e.target)) return;
 	if (!e || !e.target || shouldSkipNav(e)) return;
 	if (e.target === document) return;
@@ -171,7 +173,7 @@ function onClick (e) {
 
 
 function onDblClick (e) {
-	if (selectable !== true && selectable !== 'true') return;
+	if (!_selectable) return;
 	if (!_this.contains(e.target)) return;
 	if (shouldSkipNav(e)) return;
 
@@ -185,7 +187,7 @@ function onDblClick (e) {
 
 
 function onKeyDown (e) {
-	if (selectable !== true && selectable !== 'true') return;
+	if (!_selectable) return;
 	if (!_this.contains(e.target)) return;
 	if (shouldSkipNav(e)) return;